src/main/java/com/safetynet/alerts/service/OpsPersonService.java
package com.safetynet.alerts.service;

import java.util.ArrayList;
import java.util.List;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.safetynet.alerts.AlertsApplication;
import com.safetynet.alerts.model.PersonEntity;
import com.safetynet.alerts.repositery.PersonRepository;

/**
 * OpsPersonService is the class in charge of the person business work.
 *
 * @author Thierry Schreiner
 */
@Service
public class OpsPersonService {

    /**
     * Create a SLF4J/LOG4J LOGGER instance.
     */
    static final Logger LOGGER = LoggerFactory
            .getLogger(AlertsApplication.class);

    /**
     * OpsPersonRepository is an Interface that extends CrudRepository.
     */
    @Autowired
    private PersonRepository personRepository;

    /**
     * Class constructor - Set personRepository (IoC).
     *
     * @param pPersonRepos
     */
    public OpsPersonService(final PersonRepository pPersonRepos) {
        personRepository = pPersonRepos;
    }

    public List<String> findAllMailByCity(String pCity){
        List<PersonEntity> listPE = (List<PersonEntity>) personRepository
                .findByAddressIdCity(pCity);
System.out.println(listPE);        
        List<String> eMailList = new ArrayList<>();
        for (PersonEntity p : listPE) {
            eMailList.add(p.getEmail());
        }
        return eMailList;
    }

}
